Drop ECDHE-PSK-AES-128-GCM.
This is the best PSK cipher suite, but it's non-standard and nobody is
using it. Trivial to bring back in the future if we have need of it.

(Note that this is a no-op in Android because Android had already
disabled this cipher suite.)

(This is a cherry-pick of BoringSSL's 1feb42a2.)

(cherry picked from commit a4be71cee108bfed76ddb37552b7e48945d91b49)

Bug: 21522548
Change-Id: I2a051724500341053595f59e755349544da63ce5